chore: service code splitting
컴포넌트를 재사용하기위해 컴포넌트 기존코드에서 분리시켰습니다.
