+++ src/components/modals/Modal.tsx
@@ -1,3 +1,11 @@
+// NOTE:
+// 사용방법
+// Modal 을 import 합니다.
+// Modal 안에 내용을 입력합니다
+// ex <Modal><div><p>예시입니다.</p></div></Modal>
+// title props로 모달의 제목을 컨트롤 할 수 있습니다.
+// ex <Modal title='반달짐승의 표효'><p>어흥</p></Modal>
+
 import React, { useEffect } from "react";
 
 import closeBtn from "assets/close.svg";
@@ -37,11 +45,14 @@ export const Modal = ({ children, title }: Props): JSX.Element => {
   return (
     <>
       {/* 뒷배경 */}
-      <div onClick={onCloseModal} className="fixed top-0 bottom-0 left-0 right-0 block w-full h-full bg-[#00000040]">
+      <div
+        onClick={onCloseModal}
+        className="fixed top-0 bottom-0 left-0 right-0 block w-full h-full bg-[#00000040] z-[9999]"
+      >
         {/* 모달 */}
         <div
           onClick={stopEventBubbleing}
-          className="fixed top-[50%] left-[50%] p-6 bg-white  translate-x-[-50%] translate-y-[-50%]"
+          className="fixed top-[50%] left-[50%] p-10 bg-white  translate-x-[-50%] translate-y-[-50%]"
         >
           {/* 모달 헤더 */}
           <div className={`flex items-center justify-between pb-3 mb-8 ${titleBorder}`}>
