+++ src/components/service/data.tsx
@@ -1,24 +1,9 @@
-export const wallPaperTextureList: string[] = [
-  // "합판마루",
-  // "강화마루",
-  // "강마루",
-  // "원목마루",
-  // "하이브리드마루",
-  // "시트",
-  // "타일",
-  // "데코타일",
-  // "LVT",
-  // "마모륨",
-  "벽지",
-  "페인트",
-  "타일",
-  "포세린",
-];
+/**
+ * 여기에 벽지 카테고리가 추가되면 추가하면 됩니다.
+ */
+export const wallPaperTextureList: string[] = ["벽지", "페인트", "타일", "포세린"];
 
-export const tailTextureList: string[] = [
-  // "합지", "실크", "발포", "뮤럴", "페브릭", "방염"
-  "장판",
-  "마루",
-  "데코타일",
-  "포세린",
-];
+/**
+ * 여기에 타일 카테고리가 추가되면 추가하면 됩니다.
+ */
+export const tileTextureList: string[] = ["장판", "마루", "데코타일", "포세린"];
