+++ src/components/service/index.ts
@@ -1,3 +1,4 @@
 export * from "./ServiceItem";
 export * from "./TextureTitle";
+export * from "./InteriorSection";
 export * from "./data";
